Update to express 4.4.x, include connect
Riding on the heels of pull request #24
(https://github.com/azat-co/expressworks/pull/24), this updates the
dependencies to express 4.4.x and includes connect as a dependency
(which is required with pull request #24).

We’re seeing failures with the “what’s in query” problem when using
express 4 vs express 3.  Most new users default to use express 4, so by
including express 4 in expressworks, the problem solutions should
align.  More information can be found in this nodeschool issue:
https://github.com/nodeschool/discussions/issues/327